[BUGFIX]: OCL-Support with Multithrad worker/tasks ,#
Detail:
* Thread run with OCL. OCL start on slot_ocl_thread.
* Comm_thread_buffers are tasks check completed/incomplted scanning
* status.

ReviewedBy: R.Chatsiri
AutomergeTo:

Resolves: #
Documentation: #
Release:

	modified:   configure.ac
	modified:   src/ocl/cl/tire_parallel.cl
	modified:   src/ocl/cl_bootstrap.cpp
	modified:   src/ocl/cl_bootstrap.hpp
	modified:   src/ocl/utils/base_clutil.hpp
	modified:   src/ocl/utils/clutil_commandqueue.cpp
	modified:   src/ocl/utils/clutil_memory.cpp
	modified:   src/ocl/utils/clutil_memory.hpp
	modified:   src/scan/file_scan_policy.cpp
	modified:   src/scan/file_scan_policy.hpp
	modified:   src/test/file_policy_test.hpp
	modified:   src/test/semaphore_controller_test.cpp
	modified:   src/test/thread_sync_test.hpp
	modified:   src/threadsyncocl/buffer_sync.hpp
	modified:   src/threadsyncocl/thread_controller.cpp
	modified:   src/threadsyncocl/thread_sync.cpp
	modified:   src/threadsyncocl/thread_sync.hpp
	deleted:    src/internet/network/epoller.hpp
	deleted:    src/internet/network/gen_x509_cert.hpp
	deleted:    src/internet/pf_ring/Makefile.am
	deleted:    src/internet/pf_ring/PFRing_heandler.hpp
	deleted:    src/internet/pf_ring/pf_controller.hpp
	deleted:    src/internet/pf_ring/pf_device_manager.hpp
	deleted:    src/internet/pf_ring/pf_rules_controller.hpp
	deleted:    src/internet/sec/x509controller.hpp
	deleted:    src/test/test_suite
	deleted:    src/thread_syncocl.hpp
	deleted:    src/threadsyncocl/TimerController.hpp
	deleted:    src/threadsyncocl/monitor_controller.cpp
	deleted:    src/threadsyncocl/monitor_controller.hpp
	deleted:    src/threadsyncocl/mutex_controller.cpp
	deleted:    src/threadsyncocl/mutex_controller.hpp
	deleted:    src/threadsyncocl/thread_manager.cpp
	deleted:    src/threadsyncocl/thread_manager.hpp
	deleted:    src/utils/base/platformsocket.hpp
	deleted:    src/utils/base/util_thread.cpp
	deleted:    src/utils/base/util_thread.hpp

+++ src/ocl/cl_bootstrap.hpp
@@ -25,8 +25,6 @@ namespace hnmav_kernel
     namespace dstr = data_structure;
     namespace util = hnmav_util;
 
-    //namespace cl_bootstrap
-    //{
     template<typename UtilPlatform,
              typename TireDefine,
              typename WorkTypes,
@@ -55,7 +53,7 @@ namespace hnmav_kernel
             bool cl_process_buffer(std::vector<char>& symbol_vec_tr,
                     std::vector<size_t>&  state_vec_ptr,
                     std::vector<char>& binary_vec,
-										std::vector<uint64_t> & result_vec);
+										std::vector<uint8_t> & result_vec);
 
             // Manage command queue
             bool cl_process_commandqueue();
@@ -80,8 +78,6 @@ namespace hnmav_kernel
             TireDefine *node_tire_input;
             TireDefine *node_tire_output;
     };
-
-    // }
 
 }
